+++ releasenotes.md
@@ -3,7 +3,8 @@
 ## Version 0.10.0
 ### New Features
 - Company.CzechOrganizationNumber()
-
+- Company.FrenchSirenNumber()
+- Company.FrenchSiretNumber()
 
 ## Version 0.9.0
 Lorem is now on par with Ruby Faker for default locale ("en").